+++ model/participant_group.go
@@ -1,0 +1,8 @@
+package model
+
+// ParticipantGroup holds information about the group
+type ParticipantGroup struct {
+	ID        string `json:"id"`
+	Name      string `json:"name"`
+	ProjectID string `json:"project_id"`
+}