+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next
 
 - Add the ability to view hook events for a given subscription `hook events -s [subscription_id]`.
+- Add the ability to list participant groups via `participant list -p [project_id]`.
 
 ## 0.0.16